See if we can feature link easily from reference doc to feature doc
In theory if you add feature: to the start of a feature name we will generate
a cross link to the feature doc. This works for feature doc to config doc,
but we haven't tried it outside that context, so this is an experiment to see
if it works generally.